requirements.txt
click>=8.0.0
gitpython>=3.1.0
pyyaml>=6.0
openai>=1.0.0
groq>=0.8.0
anthropic>=0.72.0
pyperclip>=1.8.0
colorama>=0.4.6
rich>=13.0.0
tiktoken>=0.12.0
